RunningStats: uint64_t counter instead of long long

+++ src/running_stats.h
@@ -14,7 +14,7 @@ class RunningStats {
 
 	private:
 	bool na_rm;
-	unsigned long long count;
+	uint64_t count;
 	double mean, min, max, sum;
 	double M2;
 
@@ -26,7 +26,7 @@ class RunningStats {
 	
 	void reset();
 	
-	unsigned long long get_count() const;
+	double get_count() const;
 	double get_mean() const;
 	double get_min() const;
 	double get_max() const;
